@@ -6,24 +6,24 @@ publish_to: 'none'
 version: 1.0.1+2
 
 environment:
-  sdk: ">=2.16.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
-  adaptive_components: ^0.0.7
+  adaptive_components: 0.0.7
   dio: 4.0.6
   flutter:
     sdk: flutter
-  json_annotation: 4.5.0
+  json_annotation: 4.6.0
   retrofit: 3.0.1+1
-  shared_preferences: ^2.0.15
-  url_launcher: ^6.1.2
+  shared_preferences: 2.0.15
+  url_launcher: 6.1.5
 
 dev_dependencies:
-  build_runner: 2.1.11
+  build_runner: 2.2.0
   flutter_lints: 2.0.1
-  json_serializable: 6.2.0
+  json_serializable: 6.3.1
   model_generator: 5.8.1
-  retrofit_generator: 4.0.1
+  retrofit_generator: 4.0.3+1
 
 flutter:
   uses-material-design: true
