@@ -80,7 +80,7 @@ class HomePageState extends State<HomePage>
                 'An error occurred',
             style: Theme.of(context)
                 .textTheme
-                .bodyText1
+                .bodyLarge
                 ?.copyWith(color: Colors.white),
           ),
           backgroundColor: Colors.red,
